# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API server for inovantapp.",
        "title": "inovantapp API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {},
        "version": "1.0"
    },
    "host": "192.168.1.69:8080",
    "basePath": "/",
    "paths": {
...
        "/api/patients": {
            "get": {
                "description": "Get Patient list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Patient.List",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Context to return",
                        "name": "context",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter Patients by type [patiID]",
                        "name": "patiID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter Patients by type [doctID]",
                        "name": "doctID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter Patients by type [name]",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter Patients by type [email]",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter Patients by type [createdAt[gte]]",
                        "name": "createdAt[gte]",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter Patients by type [createdAt[lte]]",
                        "name": "createdAt[lte]",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.patientsListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create Patient",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Patient.Create",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Context to return",
                        "name": "context",
                        "in": "query"
                    },
                    {
                        "description": "Create new Patient",
                        "name": "Patient",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.Patient"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.patientGetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/patients/{patiID}": {
            "get": {
                "description": "Get a Patient",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Patient.Get",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Context to return",
                        "name": "context",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter Patients by type [patiID]",
                        "name": "patiID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.patientGetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update Patient",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Patient.Update",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Context to return",
                        "name": "context",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "string",
                        "description": "Patient ID",
                        "name": "patiID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Patient Update Body",
                        "name": "Patient",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.Patient"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.patientGetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.detailError": {
            "type": "object",
            "properties": {
                "domain": {
                    "type": "string"
                },
                "extendedHelp": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "locationType": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "sendReport": {
                    "type": "string"
                }
            }
        },
...
        "handler.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "$ref": "#/definitions/handler.generalError"
                }
            }
        },
...
        "handler.generalError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.detailError"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handler.patientGetResponse": {
            "type": "object",
            "properties": {
                "context": {
                    "description": "Client sets this value and server echos data in the response",
                    "type": "string"
                },
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/handler.patientResponse"
                }
            }
        },
        "handler.patientResponse": {
            "type": "object",
            "properties": {
                "item": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                }
            }
        },
        "handler.patientsListResponse": {
            "type": "object",
            "properties": {
                "context": {
                    "description": "Client sets this value and server echos data in the response",
                    "type": "string"
                },
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/handler.patientsResponse"
                }
            }
        },
        "handler.patientsResponse": {
            "type": "object",
            "properties": {
                "currentItemCount": {
                    "description": "The number of items in this result set",
                    "type": "integer",
                    "example": 1
                },
                "items": {
                    "type": "string"
                },
                "itemsPerPage": {
                    "description": "The number of items in the result",
                    "type": "integer",
                    "example": 10
                },
                "kind": {
                    "type": "string"
                },
                "lang": {
                    "description": "Indicates the language of the rest of the properties in this object (BCP 47)",
                    "type": "string",
                    "example": "pt-br"
                },
                "pageIndex": {
                    "description": "The index of the current page of items",
                    "type": "integer",
                    "example": 1
                },
                "startIndex": {
                    "description": "The index of the first item in data.items",
                    "type": "integer",
                    "example": 1
                },
                "totalItems": {
                    "description": "The total number of items available in this set",
                    "type": "integer",
                    "example": 100
                },
                "totalPages": {
                    "description": "The total number of pages in the result set.",
                    "type": "integer",
                    "example": 10
                }
            }
        },
...
        "models.Patient": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "doctID": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "info": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "patiID": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    }
}